Add unit dropdowns to choose file units
I think this is a reasonable first step but the UI of the page is
starting to break down a little - you have to select the file units
with no feedback and before you click the choose file button. Some
more work needed here which I intend to tackle in later commits.