Add test for csvExport.tableToCSV
I'm using Mocha because we're using it in Mission Patch.

I've had to replace the ES6 `export` keyword with Node's `exports` so
that the tests run. I think an alternative might've been to transpile
ES6 to CommonJS but updating the code seems reasonable.

I've had to use `textContent` instead of `innerText` as jsdom[1] doesn't
support `innerText`[2].

[1]: https://github.com/jsdom/jsdom
[2]: https://github.com/jsdom/jsdom/issues/1245