Fix bug in sprague.js
Our implementation of the Sprague interpolation was giving a different
result to the reference Matlab implementation. In the reference
implementation (doc/sprague.m) the line

      y     = a0+a1*xx+a2*xx.^2+a3*xx.^3+a4*xx.^4+a5*xx.^5;

relies on the fact that the exponentiation has a higher precedence
than the multiplication. This was not correctly implemented in
sprague.js.

The expected values in the test I've added were generated with the
following GNU Octave code:

```
input = csvread("../data/example-spectrum-10nm.csv", 2, 0)
output = sprague(input(1:16,2)', 10)'
output(1:5)
```