Script to convert CIE (1931) and CIE (1964) CSVs to JSON
Manuel provided us[1] with two CSV files containing the CIE (1931) and
CIE (1964) Colour Matching Functions that we will use to
calculate (x,y) and (x_10, y_10) values from the spectra. I've added
those CSVs and a script to convert them to JSON, which I've also
commited in the src/ directory.

[1] https://trello.com/c/hBgoFQHn/21-add-colour-calculations-to-summary-table