Add Chromaticity (CIE 1931) calculation to summary table
https://trello.com/c/hBgoFQHn/21-add-colour-calculations-to-summary-table

Manuel talked me through these calculations. We're returning (x,y,Y)
where the Y value should be the same as the Illuminance in the summary
table. I think this is the correct way to calculate `x` and `y` based
on the formula in Wikipedia[1] and those Manuel sent in Slack - but
I'm going to run them by him to check.

[1] https://en.wikipedia.org/wiki/CIE_1931_color_space#CIE_xy_chromaticity_diagram_and_the_CIE_xyY_color_space