Quote cells with commas in CSV export
If a cell value contains a ',' we have to quote that as per
RFC4180[1]. I investigated using a library to do this, but they all do
a bit too much. This works OK for now.

This became a problem with the previous commit that formats
chromaticity values as, e.g. `(0.20, 0.10, 134)`.

[1] https://tools.ietf.org/html/rfc4180