Always display two digits in exponent of scientific notation
https://trello.com/c/ObCWvm9Q/30-update-scientific-notation-format

We prefer to always display two digits in the exponent when using
scientific notation, e.g. 1.36e+02 instead of 1.36e+2

Eslint wants me to name the capture group using e.g ?<name>[1]. This
is an ECMAScript 2018 feature, and I'm not sure it improves things in
this case, so I've disabled it for this function.

[1] https://eslint.org/docs/rules/prefer-named-capture-group