Help users generate data for their research papers
There are two main use cases for Lightbox:

1. Make it possible/easy for researchers to include detailed information
about lighting conditions in their papers.

2. Allow users to visualise/explore/play with the spectral power
distribution recordings.

We've been trying to serve both use cases through a single interface but
are starting to feel the pain of that approach.

The changes in this commit add a new start point for the app
(/start.html) that aims to guide a user through generating/uploading
their data for use case 1 above.

I want to get feedback before taking this any further and this seems
like a reasonable point to stop.