Add new results page
So that I can more easily explain what one should do with the results
that are generated.

I'm only using this in the 'wizard' interface at the moment. The
original single page UI (index.html) continues to function as it did
previously.

This has introduced quite a bit of duplication. I will address that
soon.