Avoid using Bootstrap's Collapse on results page
I think the fact that we were using Bootstrap's Collapse component to
toggle the elements on the results page was a bit hidden. Modifying the
`display` style directly allows me to toggle the tables which allows me
to remove the parents containers I'd had to introduce for the Collapse
component.

A side-effect of this change is that we no longer see the animating
expansion/contraction on toggle but I think that's probably fine.