Use FileReader.readAsText() instead of Blob.text()
The `Blob.text()`[1] method isn't supported in IE, Opera, Safari and a
number of mobile browsers. Notably this meant that it wasn't working in
Safari 13 on my laptop. The `FileReader.readAsText()`[2] method appears
to be supported by all modern browsers so I'm switching to that instead.

I considered using a polyfill to provide the `Blob.text()` functionality
but went with a different native API to reduce the amount of third party
code we need.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Blob/text
[2]: https://developer.mozilla.org/en-US/docs/Web/API/FileReader/readAsText