Use babel-loader to transpile JavaScript
To increase the chances of the app working in a wider range of browsers.

I've explicitly removed IE 11 as a target browser because the app
doesn't work even after adding babel-loader[1]. Note that it was worse in
IE 11 before adding this loader but that's somewhat irrelevant given
that it still doesn't work!

Running `npx browserslist` currently lists our target browsers as:

- and_chr 81
- and_ff 68
- and_qq 10.4
- and_uc 12.12
- android 81
- baidu 7.12
- chrome 83
- chrome 81
- chrome 80
- edge 83
- edge 81
- edge 18
- firefox 76
- firefox 75
- firefox 68
- ios_saf 13.4-13.5
- ios_saf 13.3
- ios_saf 12.2-12.4
- kaios 2.5
- op_mini all
- op_mob 46
- opera 68
- opera 67
- safari 13.1
- safari 13
- samsung 11.1
- samsung 10.1

We should go through each of these, test the app and remove them if it
doesn't work. That'll give us a starting point of browsers that we do
support and we can add others back later if there's demand.

We use @babel/plugin-transform-runtime[2] to provide us with polyfills
we need rather than having to require them manually/explicitly. The
documentation for this plugin says that we need to include
@babel/runtime although I'm not 100% sure this is required - the app
certainly appeared to work locally without it.

Prior to including @babel/plugin-transform-runtime I was seeing the
following error in the JS console:

> ReferenceError: regeneratorRuntime is not defined

[1]: https://webpack.js.org/loaders/babel-loader/
[2]: https://babeljs.io/docs/en/next/babel-plugin-transform-runtime.html#technical-details