Implementation of blackbodyReferenceSpectra
This function calculates the spectra for a "Plankian radiator" at a
given colour temperature, required for step 2 of the CRI method.

The values we calculated using blackBodyReferenceSpectra disagreed by
a factor of 100 with the values in the table in step 2 of section A7.2
in Hunt and Pointer[1]. I asked Manuel about this and he said:

    At the end of paragraph 1, it says “The value of c1, however, is
    not relevant in colorimetry where only the relative spectral
    distribution of Planck’s radiation is used.” Indeed, for the CRI
    calculations, only colour, not absolute luminance or illuminance
    will be used, so it is scale independent.

    The normalisation constant is simply to normalise it, so for now,
    I’d say leave it in as is.

He's happy that the calculation is correct, so we've chosen to use
three arbitrarily chosen values from that table (divided by 100) as
our test data.

[1] Measuring Colour, Fourth Edition. R.W.G. Hunt and M.R. Pointer.