Implement equations 4.10 and 4.11 from CIE 015:2018
These equations allow us to calculate a daylight reference spectra at
a given colour temperature. This is used in the calculations of the
Colour Rendering Index when the blackbody reference spectra is not
appropriate.

Note 6 of section 4.1.2 of CIE:2018 provides a method of checking our
calculations using the values for certain specified reference
temperatures given in Table 5. Note 6 refers to truncating the
calculated value of M1 and M2 to 3 decimal places. Although this isn't
mentioned explicitly in equation 4.11 I've decided to also truncate
in the code so that the output agrees with the test data.