Implement Robertson's method for calculating CCT
This function approximates the Correlated Colour Temperature for a
given pair of (u,v) using Robertson's[1] method. Rather than go
directly to this paper, I've ported the method
`_uv_to_CCT_Robertson1968`[2] from the python `colour-science` package
and used their test data[3].

Because the calculation of CRI doesn't require the `\\Delta_{uv}`
returned by the python method, I've omitted it from this port.

It's worth noting that in the worked example in Appendix 7 of Hunt and
Pointer, their table (in step 2) gives a (u,v) of (0.3721, 0.3751)
corresponding to a T of 4224.4. When we use those values for (u,v) in
this function, we get a T of 4225.0.

I spent some time trying to track down the source of this error,
including using a different implementation of the Robertson method in
the R colorscience package[4]. This gave a different value again of
4223.8.

In the end, after discussion with Manuel, we decided to take into
account the note in CIE015:2018 (p 39):

    The computed correlated colour temperature is highly dependent on
    specifications of the computing process though the variations are
    insignificant in most practical applications... An agreement about
    details of the computing method is necessary to avoid undesirable
    discrepancies between computed correlated colour temperatures

And on page 71

    Different rounding or calculation methods can produce a difference
    of 2 K in the CCT and a difference of 1 in the colour rendering
    indices and colour fidelity index.

We're confident that our implementation of Robertson's method agrees
to within 2K of other implementations we could find.

[1] https://doi.org/10.1364/JOSA.58.001528
[2]
https://github.com/colour-science/colour/blob/8aad215d069c6b6f6dfccf78c25b8c5a97822b8c/colour/temperature/robertson1968.py#L117
[3]
https://github.com/colour-science/colour/blob/8aad215d069c6b6f6dfccf78c25b8c5a97822b8c/colour/temperature/robertson1968.py#L211
[4] https://www.rdocumentation.org/packages/colorscience/versions/1.0.8/topics/XYZ2CCT.Robertson