Calculate adaptive shift
This implements §5.7 of CIE 13.3-1995.

The test data is taken from Appendix 7 of Hunt and Pointer[1]. In
particular the spectra in step 1 of the worked calculation is
used as the test spectra. The reference spectra is a black-body
radiator with a colour temperature of 4224.09 as given in the table in
step 5 of the worked calculation.

Note that the equation to calculate `d` given in step 5 has a typo (a
- instead of a +), the equation is given correctly in §5.7 of CIE
13.3-1995.

[1] Measuring Colour, Fourth Edition. R.W.G. Hunt and M.R. Pointer (2011)