Calculate W*U*V* for reference spectra
This calculates the quantities Wri, Uri and Vri from section
5.8 (equation 5-5) of CIE 13.3:1995.

The test data is taken from the table in section 7 of Appendix 7 of
Hunt and Pointer[1]. I think there's an error in the table heading in
this book ("Test lamp" and "Reference illuminant" are swapped I think)
but I've relied on the subheadings which seem correct.

As we're only implementing the reference quantities in this commit we
can pass in the reference spectra as both inputs to the function.

[1] Measuring Colour, Fourth Edition. R.W.G. Hunt and M.R. Pointer