Add equivalent daylight alpha-opic calculations
This is the calculation of Equivalent Daylight Illuminance (EDI) /
Equivalent Daylight Luminance (EDL) as per CIE S 026/E:2018.

Section 3.7 (Note 3) of this report gives the factors used in the calculation
to 4 decimal places. We've decided to use those factors, rather than
calculate them to a higher precision using the equation at the
beginning of section 3.7.

Manual provided us with an Excel toolbox ("CIE S 026 alpha-opic
Toolbox.xlsx") from which the test data (generated from our sample
spectra) is taken. The factors are given to 4 decimal places while the
excel tool gives the results to 2 decimal places. So we've used 4
decimal places in the input to the tests and 2 in the assertions.