Calculate CRI for input spectrum
This commit threads together the calculation steps implemented in the
previous commit to calculate the CRI for any given input spectrum.

The test data is taken from Hunt and Pointer. We've relaxed the
assertion for how close the final value is to the value given in that
text, because we know that the calculated colour temperature (see
97d146b) and U*V*W values (see 2e1c6960) are slightly different to those
in the textbook.

Co-authored-by: Paul Mucur <paul@ghostcassette.com>