Add colour rendering indices to Explore view
Trello: https://trello.com/c/oLacxYGL/27-implement-standard-colour-rendering-metrics

Use the new colour rendering index calculations in the Explore view:
adding a new "Colour rendering index (CIE Ra)" row if and only if the
input spectral power distribution is at 5 nm intervals.

To do this, we now pass down the uninterpolated (but scaled) rows
through the various calculation table functions and use them to add the
new row if the conditions are right.

To be consistent with the existing rows-based calculation functions, we
add a new calculateColourRenderingIndices function that takes the
original rows and generates the colour rendering index for each
measurement, returning an array of their Ra values.

In testing this, it uncovered a bug with our calculation when the colour
temperature is over 5000 K as we were attempting to look up
D-illuminants by their wavelength * 1e-9 rather than the wavelength
directly.