Implement a linear interpolation function
This function interpolates a passed in spectrum to every 1nm. We will
use this to interpolate the daylight reference spectra and CRI test
colour spectra in cases where the user has requested the CRI calculation
for an input spectra that is not spaced every 5nm.

Trello: https://trello.com/c/oLacxYGL/27-implement-standard-colour-rendering-metrics

Co-authored-by: Paul Mucur <paul@ghostcassette.com>