Use Mocha configuration to integrate with Babel
Rather than having to explicitly require @babel/register every time we
run a test, use the Mocha configuration file to handle this for us on
every invocation of mocha.

This way, we can run individual tests or the entire test suite without
having to pass the require option to the test runner.

See https://babeljs.io/en/setup#installation and
https://github.com/mochajs/mocha-examples/tree/master/packages/babel