Use Prettier to handle all code formatting
Rather than having a custom ESLint configuration that runs only on the
src directory, use Prettier to automatically enforce a consistent code
style across the entire codebase including the tests and fixture data.

With this in place, we can relax ESLint's responsibility to non-style
issues and cut its configuration down to a minimum.

We add this to the "test" step so that every CI run will check:

* Code formatting
* Code linting
* Automated tests

See https://prettier.io