Re-implement UI with React
Trello: https://trello.com/c/MLdp5YZO/47-toggling-notation-alters-the-precision-of-values-in-the-table
Trello: https://trello.com/c/5UXomkTu/38-reported-bug-if-you-click-toggle-table-button-the-table-with-illuminance-irradiance-values-disappears-and-i-was-unable-to-get-it
Trello: https://trello.com/c/zJAtFJao/35-ensure-non-ascii-characters-display-in-other-languages

In order to resolve several issues with the UI and make the
implementation a little more idiomatic for an interactive client-side
application, use React instead of direct DOM manipulation for the
front-end.

In doing so, we've removed all other UI-related code and moved
everything into React components in a new src/javascript/components
directory.

As the style of React components were causing issues with our ESLint
configuration, we drastically simplify this to the default setup which
is far less opinionated until we introduce another style guide (e.g.
Airbnb's). This means we can now include our tests in linting as well as
the implementation. We took this opportunity to remove the
now-deprecated ESLint Loader from Webpack so that linting failures do
not prevent Webpack from compiling our code.

In order to test the new UI, we generated more realistic sample data
from the CIE 15-2014 tables[1] and replaced the existing sample files.

The new UI consolidates the existing "upload" and "Explore" sections
into one, unified results view with a toggle for the advanced
calculations. Removing the "Explore" section simplifies our Webpack
configuration as there is now only a single entrypoint into the entire
application.

In going through the new implementation, we removed any unused files,
e.g. wavelengthToRGB.

[1] http://cie.co.at/publications/colorimetry-3rd-edition

Co-authored-by: Chris Lowis <chris.lowis@gmail.com>