Fix burger nav in menu by restoring Bootstrap JS
When re-implementing the UI in React, we removed the inclusion of
Bootstrap's Collapse plugin's JavaScript thinking it was no longer used.
However, it is still used for the dropdown menu on narrow screen widths
so restore it across all HTML pages.

This means we need to process index.html and generate-csv.html with the
HTML Webpack plugin again and configure it to only include the code
needed for the collapsing menu. We also add this to the existing
configuration for the main part of the application.