Add shareable online reports
Trello: https://trello.com/c/tAY1u0fS/25-ability-to-share-urls

Use the SPDURL library to encode and decode uploads as URL-safe IDs
which we can then prompt the user to share with others when they are
viewing their results.

By having it as a separate route and component, we can restrict the
shareable report's functionality to be read-only as some options don't
make a lot of sense when a third party is accessing the report.

Co-authored-by: Chris Lowis <chris.lowis@gmail.com>