Add validation of user input CSV
Trello: https://trello.com/c/MSBr9TL7/48-validate-user-input

Before this commit we parsed the CSV and reported any errors in parsing
to the user. But we didn't enforce any constraints on the data itself.
This commit adds validation to ensure that the user uploads:

- Data with at most 5 observations (because formatting / plotting / reporting more
than that is challenging)
- Data that only contains light in the visible range (we remove any non
visible light anyway, so this helps us communicate that to the user)
- Data with a sensible wavelength column (ordered correctly, with a
constant delta and using integers that map to our look-up tables)