Use the users measurement labels everywhere
Trello: https://trello.com/c/kGLqBgRG

Before this commit we were only using the measurement labels given in
the header row of the uploaded CSV in the options for setting the
relative power. In the chart and table we were using "S0", "S1" etc
generated by sampleTitles.

With this change we parse out the header row and pass it around the
application so that we can use it where it is needed. We map it to the
SPDURL `name` property so that it is kept when a user shares a URL.

After this change `sampleTitles` is unused, so I've removed it.