Replace measurementLabels array state with object
We'd like the user to be able to change each of the measurement labels.
By making the state an object, like we do for relativePowers, it will be
easier to target the measurement label to change.

Now we have a key for each measurementLabel we can also use that in the
CalculationTableHeader.