Fix a bug in scaleSamples
Trello: https://trello.com/c/bhV5WyNj/55-order-of-magnitude-problem

When the CIE reviewed the application they found an issue with
converting from microWatts per square centimetre. We were incorrectly
scaling the data. This commit fixes that bug and improves the test
coverage to make it clearer what the calculation is doing in all cases.