Display CRI at 4 decimal places
Following some feedback from the CIE we've decided to increase the
precision of the displayed values in the calculation table.

We based our test data for the CRI calculation on Appendix 7 of Hunt and
Pointer, where the values for Ri and Ra are given to 2dp and the input
values are given at 4dp.

CIE 13.3 1995 §6.2 specificies that the values of Ri are "rounded to the
nearest whole integer". This had previously motivated our decision to do
that in the code.

The general CRI we calculate (64.13) differs slightly to the one in
Appendix 7 (64.16) for reasons that are not obvious to me. Rather than
try to track this down, I've left the tests at the same precision they
were before using `toFixed(2)`.