Add JSON file containing reference spectra
We want to allow users to chose a reference spectra to compare their
input data with. Manuel has provided us with the CSVs in this commit,
from which I've generated a JSON file with the name of the spectra as
the key and then a hash of { wavelength => value } for each.

Some of the spectra are given in 5nm spacing, others in 1nm spacing. We
intend to use the Sprague interpolation routine in the code base to
interpolate before displaying and we can do that on-the-fly at least in
the first instance, so I've just copied the values as-is into the JSON.