Add a select box to the chart to add a reference
Chart.js handles plotting arbitrary x,y points on the plot, so instead
of interpolating our reference spectra data, we can just plot the points
we have directly. I've had to reformat the data in `rows` to match this
format, and change the chart type to "linear".

I've added the ability to scale the reference to log10, and normalised
units, but the spectra themselves are in relative units. We'll scale
them to a specific lux value next.