Render content for about page from a markdown file
This proved to a be a bit tricky and I'm still not sure this is the best
way to do it. The combination of the html-loader & markdown-loader mean
that the About JSX component has access to the HTML rendered from the
About.md markdown file. However, by default this HTML is escaped when
rendered within the component and I've used the react-html-parser
package to render the HTML in its raw unescaped form.

Another option might be to render the whole page as static HTML, but I'm
not sure how well that plays with using the React router.