Style tables using bootstrap
This applies bootstrap classes to all tables across the site. My
motivation for doing this is to style the table on the About page to
make it a bit more like how GitHub renders the table. I couldn't
immediately find a way to add CSS classes to the table in its raw
markdown form and so this was the only way I could see to apply the
styles.

Note that there are tables on other pages which are affected by this
style change and that might not be desirable.