Change active state of navbar items depending on current path
I've achieved this by extracting the existing "static" bootstrap navbar
component [1] into a custom react component wrapped by withRouter [2]
from react-router which means we can get access to the current path.

While making these changes, I decided to change the "Report light
exposure" link to take the user to the "/generate-csv" path. I'm not
sure this is definitely correct, but I don't think it's made things any
worse and it makes it easier to see the nav-item active state being
changed.

I've also added a dynamic version of the sr-only span element to each
nav-item and the nav-brand; I assume this is to improve accessibility.

Note that the nav-brand "luox" looks as if it's always active. I'm not
sure how to address that, but it might help if we add more
nav-items; and/or add a logo or make the "luox" text bigger by making it
a heading.

Other changes to the nav bar HTML were forced either by JSX or prettier
requirements.

[1]: https://getbootstrap.com/docs/4.4/components/navbar/
[2]: https://reactrouter.com/web/api/withRouter