Allow CSV header row to be optional
Trello: https://trello.com/c/f2XMS14v

If the first row of the CSV contains only numeric values, we assume
there is no header row, define appropriate header values and use the
first row of the CSV as a data row rather than the header row.

Note that I've effectively reverted the validation rule about header
values needed to start with a letter. This is partly because I belatedly
realised numeric values will be instances of Number at this point rather
than String and so testing against a regular expression doesn't make
sense. Also now we have the new header logic, I don't think this
validation rule makes sense any more.

A possible enhancement to this behaviour would be to display a warning
message to indicate that we have assumed the first row is a row of data.