Prevent negative lux for relative spectra
Trello: https://trello.com/c/TCC9Pw4Y

I'm not sure it's a perfect solution, but I've set some more properties
on the number input and triggered HTML5 validation on the "input" event
so that the built in browser validation error messages are displayed. We
have to add custom code to trigger the validation, because we're not
submitting the form.

* The "step" property was defaulting to 1, which meant that non-integers
were not considered valid. I've fixed this by setting the "step"
property to "any".

* I've set the "min" property to 0.00001 based on this comment [1] from
Manuel. It means the validation message is potentially slightly
confusing and it might make sense to override this to say "Value must be
greater than 0" rather than "Value must be greater or equal to 0.00001".

* I've set the "required" property, because we expect a value to be
provided and it is possible to delete all the characters in the input
field.

* I've also added a CSS style to highlight invalid input fields when you
change the focus to another field.

On the "backend", I've changed the code in the useEffect function that
was already only accepting numeric values so that it now also only
accepts positive numbers. Note that I was initially a bit confused that
it is code in the relativeToAbsolute() function (in calculations.js)
which defaults missing values (and 0) to 1.

[1]: https://trello.com/c/TCC9Pw4Y/70-prevent-negative-lux-for-relative-spectra#comment-5ffb145635e6a53f24e0df50