Update filename displayed when CSV file is uploaded
Previously we were using the custom file input from bootstrap which was
introduced in this commit [1]. However, we hadn't wired up the relevant
JS to set the filename when a file was selected.

Rather than wire up the relevant JS, I've decided to effectively revert
the original commit and use the standard bootstrap file input. I'm not
convinced that the large width of the custom file input looked very good
aesthetically, at least not on a desktop display. This means the
selected filename is now displayed when a file is selected.

[1]: ff167e6e42e987bec5aa02a94b0b84e54464a2fb