WIP: Various fixes for uvToCorrelatedColourTemperatureRobertson function
There were a couple of mistakes in the port from this Python code [1].

* `for i in range(1, 31)` in Python is equivalent to `for (let i = 1; i
<= 30; i += 1)` in JavaScript not `for (let i = 1; i <= 31; i += 1)`

* The keys in robertson.json should start at 0 not 1 to match the
ISOTEMPERATURE_LINES_ROBERTSON1968 array in the Python code. I've also
updated the robertson-csv-to-json.rb script which is used to generate
the JSON file from a CSV source file.

* We need to give wrRuvtPrevious a default value, so that accessing
properties on it will not raise a TypeError.

* I've added a test to ensure a TypeError exception is not raised in a
bit if an edge case that Manuel discovered. This test is now passing.

* I've added a test related to the discrepancy mentioned in this
commit [2]. The calculated CCT value still doesn't match the value in
the worked example in Appendix 7.2 of the Hunt & Pointer book, but it's
pretty close and it *does* exactly match the CCT calculated using the
Python colour-science package.

[1]: https://github.com/colour-science/colour/blob/8aad215d069c6b6f6dfccf78c25b8c5a97822b8c/colour/temperature/robertson1968.py#L117-L184
[2]: 28e9354c17d539e4a27b925e1dd3d80f5fc40d26