Do not interpolate input data when applying action functions
Trello: https://trello.com/c/aynuTTR7

Previously we were interpolating the input data at 1nm intervals so that
we could guarantee finding a corresponding value in the action function
reference data. However, this interpolation led to inconsistencies with
values that @spitschan calculated in Matlab and obtained from the CIE
toolbox. Removing the call to interpolationData() seems to bring the
values into line with the figures calculated by @spitschan.

|             | Matlab (W⋅m⁻²)     | CIE toolbox (W⋅m⁻²) | Luox (before) (mW⋅m⁻²) | Luox (after) (mW⋅m⁻²)    |
|-------------|--------------------|---------------------|------------------------|--------------------------|
| S-cone-opic |   779.3089768492   |              779.31 |           779,686.8913 |          779,308.9768 ✅ |
| M-cone-opic | 1,429.290778274974 |            1,429.29 |         1,429,273.9637 |        1,429,290.7783 ✅ |
| L-cone-opic | 1,610.210532773205 |            1,610.21 |         1,610,207.2168 |        1,610,210.5328 ✅ |
| Rhodopic    | 1,295.841683222725 |            1,295.84 |         1,295,849.9849 |        1,295,841.6832 ✅ |
| Melanopic   | 1,152.778699183284 |            1,152.78 |         1,152,942.4947 |        1,152,778.6992 ✅ |

At @spitschan's suggestion we've recently restricted the input data to
have wavelength intervals of between 1nm and 10nm. The following action
function reference data seem to have data at 1nm intervals. I think
these are the only reference data needed in these "opic" calculations
and so this change should be OK.

* src/data/ciexyz64.json
* src/data/ciexyz31.json
* src/data/cies026.json
* src/data/vl1924.json

Note that I've added some "integration" tests of sorts which check the
calculate() function returns the expected "opic" values match the values
calculated by @spitschan using Matlab and the CIE toolbox. The first
test checks the figures to more decimal places than the other tests,
because @spitschan supplied the expected results to that accuracy.

The calculateColourRenderingIndex() calculation uses a different approach to
wavelength sampling and should not be affected by this change.

The luminant calculation will have been affected by this removal of
interpolation, but I plan to check that separately.