Correct multiplicative constant, K, for illuminance calculation
Trello: https://trello.com/c/aynuTTR7

This changes the value of K from 683 to 683.002. The latter is the value
that @spitschan thinks is the correct value despite the CIE toolbox
using 683.0015478. @spitschan provided this quote:

> The lumen is defined to be unity for a radiant energy of 1/683 W at a
> frequency of 540 THz, which corresponds to a standard air wavelength of
> 555.016 nm rather than 555 nm, which is the peak of the luminosity
> curve. The value of y(λ) is 0.999997 at 555.016 nm, so that a value of
> 683/0.999997 = 683.002 is the multiplicative constant.

For further details see the conversations in the lightbox Slack channel
on Fri 15 Jan 2021, e.g. this message [1].

[1]: https://gofreerange.slack.com/archives/CRLTY4Z24/p1610726311031000