Make sharing link more discoverable
Trello: https://trello.com/c/Az52kTKS

* Initially only a maximum of 5 rows of the full results table in step 5
are displayed.

* A row displaying ellipsis is added to the table to indicate that other
rows are missing.

* A "Display all rows?" checkbox is added at the top of that table. When
checked all rows in the table are displayed.

* A "Copy to clipboard" primary button replaces the muted form text.

* Clicking the button copies the link to the clipboard, the button is
disabled, and the button text changes to "Copied!" to give the user
feedback.

* After 1 second the button text changes back to "Copy to clipboard" and
the button is re-enabled.

Note that it might be better to extract the button into its own React
component, but this will do for now.