Run the prettier tool to fix build errors
I ran the following command:

    prettier -w .

While these changes to the markdown document aren't essential, I think
they are on the whole an improvement and so I'm going to stick with
them.

However, if we want to make it easier for @spitschan to make this kind
og change, it might make sense to exclude this markdown file from the
prettier linting.