Display version info derived from git at build time
I've elected to display both the latest git tag *and* the latest commit
SHA, because the latter is guaranteed to allow us to recreate the state
of the codel whereas the former is more of a human-friendly version
identifier which has to be manually added.

It's not ideal that the "Version information" heading level and DOM ID
depend on the theoretically more variable table of content and headings
in the About.md markdown. However, it'll do as a starting point. If we
run into problems, we can probably use some kind of
templating/substitution in the About.md file.

Note that this introduces a build-time dependency on bash to run the
shell script which generates the version.json file, but I don't think
that's a big deal.